#1811: Refactored code to decentralize config types.
