+++ internal/core/command/config/config.go
@@ -15,8 +15,6 @@
 package config
 
 import (
-	"github.com/edgexfoundry/edgex-go/internal/pkg/config"
-
 	"github.com/edgexfoundry/go-mod-bootstrap/bootstrap/interfaces"
 	bootstrapConfig "github.com/edgexfoundry/go-mod-bootstrap/config"
 )
@@ -25,7 +23,7 @@ import (
 type ConfigurationStruct struct {
 	Writable    WritableInfo
 	Clients     map[string]bootstrapConfig.ClientInfo
-	Databases   config.DatabaseInfo
+	Databases   map[string]bootstrapConfig.Database
 	Logging     bootstrapConfig.LoggingInfo
 	Registry    bootstrapConfig.RegistryInfo
 	Service     bootstrapConfig.ServiceInfo
@@ -94,6 +92,6 @@ func (c *ConfigurationStruct) GetRegistryInfo() bootstrapConfig.RegistryInfo {
 }
 
 // GetDatabaseInfo returns a database information map.
-func (c *ConfigurationStruct) GetDatabaseInfo() config.DatabaseInfo {
+func (c *ConfigurationStruct) GetDatabaseInfo() map[string]bootstrapConfig.Database {
 	return c.Databases
 }
