+++ internal/core/metadata/config/config.go
@@ -25,7 +25,7 @@ import (
 type ConfigurationStruct struct {
 	Writable      WritableInfo
 	Clients       map[string]bootstrapConfig.ClientInfo
-	Databases     config.DatabaseInfo
+	Databases     map[string]bootstrapConfig.Database
 	Logging       bootstrapConfig.LoggingInfo
 	Notifications config.NotificationInfo
 	Registry      bootstrapConfig.RegistryInfo
@@ -96,6 +96,6 @@ func (c *ConfigurationStruct) GetRegistryInfo() bootstrapConfig.RegistryInfo {
 }
 
 // GetDatabaseInfo returns a database information map.
-func (c *ConfigurationStruct) GetDatabaseInfo() config.DatabaseInfo {
+func (c *ConfigurationStruct) GetDatabaseInfo() map[string]bootstrapConfig.Database {
 	return c.Databases
 }
