+++ internal/security/proxy/config/config.go
@@ -19,8 +19,6 @@ import (
 	"fmt"
 	"net/url"
 
-	"github.com/edgexfoundry/edgex-go/internal/pkg/config"
-
 	"github.com/edgexfoundry/go-mod-bootstrap/bootstrap/interfaces"
 	bootstrapConfig "github.com/edgexfoundry/go-mod-bootstrap/config"
 	"github.com/edgexfoundry/go-mod-secrets/pkg/providers/vault"
@@ -165,6 +163,6 @@ func (c *ConfigurationStruct) GetRegistryInfo() bootstrapConfig.RegistryInfo {
 }
 
 // GetDatabaseInfo returns a database information map.
-func (c *ConfigurationStruct) GetDatabaseInfo() config.DatabaseInfo {
+func (c *ConfigurationStruct) GetDatabaseInfo() map[string]bootstrapConfig.Database {
 	panic("GetDatabaseInfo() called unexpectedly.")
 }
